Simple updates
Added a where clause to only show the list at are public on the
publiclist.php page. Added a different Database link to the lists,users
php. while taking away the db file from publiclist.php since it was
being used twice. Added a default 0 to the price to prevent errors for
now.